+++ src/screenComponents/missileTubeControls.cpp
@@ -1,21 +1,23 @@
-#include "playerInfo.h"
-#include "i18n.h"
 #include "missileTubeControls.h"
+#include "i18n.h"
+#include "playerInfo.h"
 #include "powerDamageIndicator.h"
-#include "components/warpdrive.h"
+
+#include "components/collision.h"
 #include "components/missiletubes.h"
 #include "components/missileWeaponTarget.h"
-#include "components/collision.h"
 #include "components/target.h"
+#include "components/warpdrive.h"
+
 #include "systems/missilesystem.h"
 
 #include "gui/gui2_button.h"
-#include "gui/gui2_progressbar.h"
 #include "gui/gui2_label.h"
+#include "gui/gui2_progressbar.h"
 #include "gui/gui2_togglebutton.h"
 
 GuiMissileTubeControls::GuiMissileTubeControls(GuiContainer* owner, string id)
-: GuiElement(owner, id), load_type(MW_None), manual_aim(false), missile_target_angle(0)
+: GuiElement(owner, id)
 {
     setSize(GuiElement::GuiSizeMax, GuiElement::GuiSizeMax);
     setAttribute("layout", "verticalbottom");
@@ -26,35 +28,40 @@ GuiMissileTubeControls::GuiMissileTubeControls(GuiContainer* owner, string id)
     for (int n = MW_MaxTypes-1; n >= 0; n--)
     {
         load_type_rows[n].layout = new GuiElement(this, id + "_ROW_" + string(n));
-        load_type_rows[n].layout->setSize(GuiElement::GuiSizeMax, 40)->setAttribute("layout", "horizontal");
+        load_type_rows[n].layout
+            ->setSize(GuiElement::GuiSizeMax, 40.0f)
+            ->setAttribute("layout", "horizontal");
 
-        load_type_rows[n].button = new GuiToggleButton(load_type_rows[n].layout, id + "_MW_" + string(n), MissileWeaponDataRegistry::instance().getNameForIndex(n), [this, n](bool value) {
-            if (value)
-                load_type = n;
-            else
-                load_type = MW_None;
-            for(int idx = 0; idx < MW_MaxTypes; idx++)
-                load_type_rows[idx].button->setValue(idx == load_type);
-        });
-        load_type_rows[n].button->setTextSize(28)->setSize(200, 40);
+        load_type_rows[n].button = new GuiToggleButton(load_type_rows[n].layout, id + "_MW_" + string(n), MissileWeaponDataRegistry::instance().getNameForIndex(n),
+            [this, n](bool value)
+            {
+                if (value) load_type = n;
+                else load_type = MW_None;
+
+                for (int idx = 0; idx < MW_MaxTypes; idx++)
+                    load_type_rows[idx].button->setValue(idx == load_type);
+            }
+        );
+        load_type_rows[n].button
+            ->setTextSize(28.0f)
+            ->setSize(200.0f, 40.0f);
     }
-    load_type_rows[0].button->setIcon("gui/icons/weapon-homing.png");
-    load_type_rows[2].button->setIcon("gui/icons/weapon-mine.png");
-    load_type_rows[3].button->setIcon("gui/icons/weapon-emp.png");
-    load_type_rows[1].button->setIcon("gui/icons/weapon-nuke.png");
-    load_type_rows[4].button->setIcon("gui/icons/weapon-hvli.png");
 }
 
 static string getTubeName(float direction)
 {
     if (std::abs(angleDifference(0.0f, direction)) <= 45)
-        return tr("tube","Front");
+        return tr("tube", "Front");
+
     if (std::abs(angleDifference(90.0f, direction)) < 45)
-        return tr("tube","Right");
+        return tr("tube", "Right");
+
     if (std::abs(angleDifference(-90.0f, direction)) < 45)
-        return tr("tube","Left");
+        return tr("tube", "Left");
+
     if (std::abs(angleDifference(180.0f, direction)) <= 45)
-        return tr("tube","Rear");
+        return tr("tube", "Rear");
+
     return "?" + string(direction);
 }
 
@@ -76,8 +83,12 @@ void GuiMissileTubeControls::onUpdate()
 
     for (int n = 0; n < MW_MaxTypes; n++)
     {
-        load_type_rows[n].button->setText(MissileWeaponDataRegistry::instance().getNameForIndex(n) + " [" + string(tubes->storage[n]) + "/" + string(tubes->storage_max[n]) + "]");
+        auto& registry = MissileWeaponDataRegistry::instance();
+        load_type_rows[n].button->setText(registry.getNameForIndex(n) + " [" + string(tubes->storage[n]) + "/" + string(tubes->storage_max[n]) + "]");
         load_type_rows[n].layout->setVisible(tubes->storage_max[n] > 0);
+
+        auto icon = registry.getIcon(n);
+        if (!icon.empty()) load_type_rows[n].button->setIcon(icon);
     }
 
     for (unsigned int n = 0; n < tubes->mounts.size(); n++)
@@ -85,26 +96,39 @@ void GuiMissileTubeControls::onUpdate()
         if (rows.size() <= n) createTubeRow();
         auto& tube = tubes->mounts[n];
         rows[n].layout->show();
-        if (tube.canOnlyLoad(2))
-            rows[n].fire_button->setIcon("gui/icons/weapon-mine", sp::Alignment::CenterLeft);
+        auto& registry = MissileWeaponDataRegistry::instance();
+
+        if (tube.type_loaded >= 0)
+        {
+            auto icon = registry.getIcon(tube.type_loaded);
+            if (!icon.empty())
+                rows[n].fire_button->setIcon(icon, sp::Alignment::CenterLeft);
+            else
+                rows[n].fire_button->setIcon("gui/icons/missile", sp::Alignment::CenterLeft, tube.direction);
+        }
         else
             rows[n].fire_button->setIcon("gui/icons/missile", sp::Alignment::CenterLeft, tube.direction);
 
         switch (tube.state)
         {
         case MissileTubes::MountPoint::State::Empty:
-            rows[n].load_button->setEnable(tube.canLoad(load_type));
-            rows[n].load_button->setText(tr("missile","Load"));
+            rows[n].load_button
+                ->setText(tr("missile", "Load"))
+                ->setEnable(tube.canLoad(load_type));
 
             if (health <= 0) rows[n].load_button->disable();
 
-            rows[n].fire_button->disable()->show();
-            rows[n].fire_button->setText(getTubeName(tube.direction) + ": " + tr("missile","Empty"));
+            rows[n].fire_button
+                ->setText(getTubeName(tube.direction) + ": " + tr("missile", "Empty"))
+                ->disable()
+                ->show();
+
             rows[n].loading_bar->hide();
             break;
         case MissileTubes::MountPoint::State::Loaded:
             rows[n].load_button->setText(tr("missile","Unload"));
-            if ((health <= 0) || (power_level <=0))
+
+            if ((health <= 0) || (power_level <= 0))
             {
                 rows[n].fire_button
                     ->disable()
@@ -122,32 +146,50 @@ void GuiMissileTubeControls::onUpdate()
                     ->enable()
                     ->show();
             }
+
             rows[n].fire_button->setText(getTubeName(tube.direction) + ": " + MissileWeaponDataRegistry::instance().getNameForIndex(tube.type_loaded));
             rows[n].loading_bar->hide();
             break;
         case MissileTubes::MountPoint::State::Loading:
-            rows[n].load_button->disable();
-            rows[n].load_button->setText(tr("missile","Load"));
-            rows[n].fire_button->hide();
-            rows[n].fire_button->setText(getTubeName(tube.direction) + ": " + MissileWeaponDataRegistry::instance().getNameForIndex(tube.type_loaded));
-            rows[n].loading_bar->show();
-            rows[n].loading_bar->setValue(1.0f - tube.delay / tube.load_time);
-            rows[n].loading_label->setText(tr("missile","Loading"));
+            rows[n].load_button
+                ->setText(tr("missile", "Load"))
+                ->disable();
+
+            rows[n].fire_button
+                ->setText(getTubeName(tube.direction) + ": " + MissileWeaponDataRegistry::instance().getNameForIndex(tube.type_loaded))
+                ->hide();
+
+            rows[n].loading_bar
+                ->setValue(1.0f - tube.delay / tube.load_time)
+                ->show();
+
+            rows[n].loading_label->setText(tr("missile", "Loading"));
             break;
         case MissileTubes::MountPoint::State::Unloading:
-            rows[n].load_button->disable();
-            rows[n].load_button->setText(tr("missile","Unload"));
-            rows[n].fire_button->hide();
-            rows[n].fire_button->setText(MissileWeaponDataRegistry::instance().getNameForIndex(tube.type_loaded));
-            rows[n].loading_bar->show();
-            rows[n].loading_bar->setValue(tube.delay / tube.load_time);
-            rows[n].loading_label->setText(tr("missile","Unloading"));
+            rows[n].load_button
+                ->setText(tr("missile", "Unload"))
+                ->disable();
+
+            rows[n].fire_button
+                ->setText(MissileWeaponDataRegistry::instance().getNameForIndex(tube.type_loaded))
+                ->hide();
+
+            rows[n].loading_bar
+                ->setValue(tube.delay / tube.load_time)
+                ->show();
+
+            rows[n].loading_label->setText(tr("missile", "Unloading"));
             break;
         case MissileTubes::MountPoint::State::Firing:
-            rows[n].load_button->disable();
-            rows[n].load_button->setText(tr("missile","Load"));
-            rows[n].fire_button->disable()->show();
-            rows[n].fire_button->setText(tr("missile","Firing"));
+            rows[n].load_button
+                ->setText(tr("missile", "Load"))
+                ->disable();
+
+            rows[n].fire_button
+                ->setText(tr("missile", "Firing"))
+                ->disable()
+                ->show();
+
             rows[n].loading_bar->hide();
         }
 
@@ -158,6 +200,8 @@ void GuiMissileTubeControls::onUpdate()
     while (rows.size() > tubes->mounts.size()) removeTubeRow();
 
     // Handle hotkeys for default missile types.
+    // TODO: Either make these hotkeys match by missile type string, or make
+    // them select by index instead of missile type.
     if (keys.weapons_select_homing.getDown()) selectMissileWeapon(0);
     if (keys.weapons_select_nuke.getDown()) selectMissileWeapon(1);
     if (keys.weapons_select_mine.getDown()) selectMissileWeapon(2);
@@ -180,8 +224,10 @@ void GuiMissileTubeControls::onUpdate()
                     target_entity = tgt->entity;
                 else if (auto tgt = my_spaceship.getComponent<Target>())
                     target_entity = tgt->entity;
+
                 target_angle = MissileSystem::calculateFiringSolution(my_spaceship, tubes->mounts[n], target_entity);
-                if (target_angle == std::numeric_limits<float>::infinity()) {
+                if (target_angle == std::numeric_limits<float>::infinity())
+                {
                     auto transform = my_spaceship.getComponent<sp::Transform>();
                     target_angle = (transform ? transform->getRotation() : 0.0f) + tubes->mounts[n].direction;
                 }
@@ -218,50 +264,68 @@ void GuiMissileTubeControls::createTubeRow()
     TubeRow row;
     row.layout = new GuiElement(tube_rows_layout, id + "_ROW_" + string(n));
     row.layout->setAttribute("layout", "horizontal");
-    row.load_button = new GuiButton(row.layout, id + "_" + string(n) + "_LOAD_BUTTON", "Load", [this, n]() {
-        if (!my_spaceship) return;
-        auto tubes = my_spaceship.getComponent<MissileTubes>();
-        if (!tubes) return;
-        if (tubes->mounts[n].state == MissileTubes::MountPoint::State::Empty)
+
+    row.load_button = new GuiButton(row.layout, id + "_" + string(n) + "_LOAD_BUTTON", tr("Load"),
+        [this, n]()
         {
-            if (load_type != MW_None)
+            if (!my_spaceship) return;
+
+            auto tubes = my_spaceship.getComponent<MissileTubes>();
+            if (!tubes) return;
+
+            if (tubes->mounts[n].state == MissileTubes::MountPoint::State::Empty)
             {
-                my_player_info->commandLoadTube(n, load_type);
+                if (load_type != MW_None)
+                    my_player_info->commandLoadTube(n, load_type);
             }
+            else my_player_info->commandUnloadTube(n);
         }
-        else
+    );
+    row.load_button->setSize(130.0f, GuiElement::GuiSizeRow);
+
+    row.fire_button = new GuiButton(row.layout, id + "_" + string(n) + "_FIRE_BUTTON", tr("Fire"),
+        [this, n]()
         {
-            my_player_info->commandUnloadTube(n);
-        }
-    });
-    row.load_button->setSize(130, 50);
-    row.fire_button = new GuiButton(row.layout, id + "_" + string(n) + "_FIRE_BUTTON", "Fire", [this, n]() {
-        if (!my_spaceship) return;
-        auto tubes = my_spaceship.getComponent<MissileTubes>();
-        if (!tubes) return;
-        if (tubes->mounts[n].state == MissileTubes::MountPoint::State::Loaded)
-        {
-            float target_angle = missile_target_angle;
-            if (!manual_aim) {
-                sp::ecs::Entity target_entity;
-                if (auto tgt = my_spaceship.getComponent<MissileWeaponTarget>())
-                    target_entity = tgt->entity;
-                else if (auto tgt = my_spaceship.getComponent<Target>())
-                    target_entity = tgt->entity;
-                target_angle = MissileSystem::calculateFiringSolution(my_spaceship, tubes->mounts[n], target_entity);
-                if (target_angle == std::numeric_limits<float>::infinity()) {
-                    auto transform = my_spaceship.getComponent<sp::Transform>();
-                    target_angle = (transform ? transform->getRotation() : 0.0f) + tubes->mounts[n].direction;
+            if (!my_spaceship) return;
+
+            auto tubes = my_spaceship.getComponent<MissileTubes>();
+            if (!tubes) return;
+
+            if (tubes->mounts[n].state == MissileTubes::MountPoint::State::Loaded)
+            {
+                float target_angle = missile_target_angle;
+
+                if (!manual_aim)
+                {
+                    sp::ecs::Entity target_entity;
+
+                    if (auto tgt = my_spaceship.getComponent<MissileWeaponTarget>())
+                        target_entity = tgt->entity;
+                    else if (auto tgt = my_spaceship.getComponent<Target>())
+                        target_entity = tgt->entity;
+
+                    target_angle = MissileSystem::calculateFiringSolution(my_spaceship, tubes->mounts[n], target_entity);
+                    if (target_angle == std::numeric_limits<float>::infinity())
+                    {
+                        auto transform = my_spaceship.getComponent<sp::Transform>();
+                        target_angle = (transform ? transform->getRotation() : 0.0f) + tubes->mounts[n].direction;
+                    }
                 }
+                my_player_info->commandFireTube(n, target_angle);
             }
-            my_player_info->commandFireTube(n, target_angle);
         }
-    });
-    row.fire_button->setSize(200, 50);
-    (new GuiPowerDamageIndicator(row.load_button, id + "_" + string(n) + "_PDI", ShipSystem::Type::MissileSystem, sp::Alignment::CenterRight))->setSize(GuiElement::GuiSizeMax, GuiElement::GuiSizeMax);
+    );
+    row.fire_button->setSize(200.0f, GuiElement::GuiSizeRow);
+
+    (new GuiPowerDamageIndicator(row.load_button, id + "_" + string(n) + "_PDI", ShipSystem::Type::MissileSystem, sp::Alignment::CenterRight))
+        ->setSize(GuiElement::GuiSizeMax, GuiElement::GuiSizeMax);
+
     row.loading_bar = new GuiProgressbar(row.layout, id + "_" + string(n) + "_PROGRESS", 0, 1.0, 0);
-    row.loading_bar->setColor(glm::u8vec4(128, 128, 128, 255))->setSize(200, 50);
-    row.loading_label = new GuiLabel(row.loading_bar, id + "_" + string(n) + "_PROGRESS_LABEL", "Loading", 35);
+    row.loading_bar
+        ->setColor(glm::u8vec4(128, 128, 128, 255))
+        ->setSize(200.0f, GuiElement::GuiSizeRow);
+
+    row.loading_label = new GuiLabel(row.loading_bar, id + "_" + string(n) + "_PROGRESS_LABEL", "Loading", 35.0f);
     row.loading_label->setSize(GuiElement::GuiSizeMax, GuiElement::GuiSizeMax);
 
     rows.push_back(row);
@@ -276,6 +340,7 @@ void GuiMissileTubeControls::removeTubeRow()
 void GuiMissileTubeControls::selectMissileWeapon(int type)
 {
     load_type = type;
-    for(int idx = 0; idx < MW_MaxTypes; idx++)
+
+    for (int idx = 0; idx < MW_MaxTypes; idx++)
         load_type_rows[idx].button->setValue(idx == type);
 }